Fix failing tests from 3bb4210b, defaulting accept to nil.
When handling of the accept header was defaulted to nil, rather than the first
content type, the tests were not fully updated to reflect the change.

Now, as expected, an accept header that does not match any provided content
type will result in no valid response and, consequently, an error back to the
client.

This correctly mirrors WebMachine behaviour.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>