Add `defhandler` to underlay `defsimplehandler`...
`defhandler` wraps the complexity of building a handler, together with default
behaviour, in a simple macro that expands, more or less, to define a type,
augment the default behaviours with the supplied behaviours, and then
creates a ready-to-use instance for the user.

`defsimplehandler` is reimplemented in terms of `defhandler`, and the complex
server test is also implemented using the wrapper macro.

Finally, update the documentation to reflect these changes.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>